Fixed failing on a special case
SwipeSensor would fail to detect a line if two adjacent points were
identical. This is because it assumed that the normalized difference
between each adjacent points would have a magnitude of 1, but did not
account for when two adjacent points are identical, in which case the
magnitude is 0. This has been fixed with an option to ignore zero
vectors.