Applied a basic touch area
The y position of a touch is checked to see whether it falls within the
GUI or not (by that I mean the black bar at the top of the screen). All
touches that move the ball must be below this black bar. The code within
GameControls.cs in the class TouchInput controls this. It uses the hacky
magic number of 2f to represent the height of the black bar (measuring
down 2f from the very top of the screen). The position of the top of the
screen is calculated by copy-pasting the code from GUIAnchor.cs which
was used to align the GUI elements to the right corner of the screen.