Level buttons now clickable
There is however a bug that does not show up when testing on the
computer. It occurs if you click a button after already having clicked a
button. The button you previously clicked gets clicked again. This is
presumably down to the messy way in which the hit event is passed from
the controller down to the button, combined with the physical simulation
running at a slower pace on mobile devices.