Added brick walls and flowerbeds
Many new art assets were drawn for the walls and flowerbeds. These were
then placed in a new atlas, LawnAtlas, which is intended for all the
lawn related sprites, and so other existing sprites should be moved
there. In order to generate the physics mesh for the curved wall, the
GenerateMesh*.js classes were imported from the old Domino games. Unlike
the rest of this project, which is C#, the GenerateMesh code is written
in JavaScript. There is currently no point in rewriting the code in C#
since it is modular and works just as needed.